requirements.txt
agora_token_builder==1.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.10
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autobahn==25.10.2
Automat==25.4.16
bcrypt==4.3.0
billiard==4.2.1
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.4.0
certifi==2025.7.14
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.5
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
diff-match-patch==20241021
distro==1.9.0
Django==5.2.6
django-appconf==1.1.0
django-celery-beat==2.8.1
django-ckeditor-5==0.2.15
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-cuser==2017.3.16
django-filter==24.3
django-import-export==4.3.10
django-jazzmin==3.0.1
django-modeltranslation==0.19.16
django-nested-admin==4.1.6
django-parler==2.3
django-select2==8.1.2
django-timezone-field==7.1
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
drf-spectacular==0.29.0
drf-yasg==1.21.8
ecdsa==0.19.1
elevenlabs==2.7.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
firebase-admin==6.5.0
fonttools==4.60.1
frozenlist==1.8.0
git-filter-repo==2.47.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
langchain-core==0.3.69
langchain-groq==0.3.6
langsmith==0.4.8
Mako==1.3.10
MarkupSafe==3.0.2
modeltranslation==0.25
msgpack==1.1.1
multidict==6.7.0
numpy==2.3.5
openpyxl==3.1.2
orjson==3.11.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.4.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-monkey-business==1.1.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.37.0
reportlab==4.0.4
requests==2.32.4
requests-toolbelt==1.0.0
resend==2.11.0
rpds-py==0.30.0
rsa==4.9.1
sendgrid==6.12.4
service-identity==24.2.0
setuptools==80.9.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
tablib==3.8.0
tenacity==9.1.2
tinycss2==1.4.0
twilio==9.3.0
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
zope.interface==8.0
zopfli==0.2.3.post1
zstandard==0.23.0
